Mario

+++ Mario/Mario/Beetle.cpp
@@ -3,8 +3,8 @@
 
 Beetle::Beetle(int xMinion, int yMinion, bool moveDirection)
 {
-	this->xMinion = xMinion;
-	this->yMinion = yMinion;
+	this->xMinion = (float)xMinion;
+	this->yMinion = (float)yMinion;
 	this->minionState = 0;
 	this->id = Window::GetMap()->GetMapType() == Window::GetMap()->BanNgay || Window::GetMap()->GetMapType() == Window::GetMap()->HoangHon || Window::GetMap()->GetMapType() == Window::GetMap()->BuoiToi || Window::GetMap()->GetMapType() == Window::GetMap()->BinhMinh ? 53 : Window::GetMap()->GetMapType() == Window::GetMap()->LongDat ? 55 : 57;
 	this->moveDirection = moveDirection;
