+++ Mario/Mario/Bowser.cpp
@@ -3,8 +3,8 @@
 
 Bowser::Bowser(int xMinion, int yMinion)
 {
-	this->xMinion = xMinion;
-	this->yMinion = yMinion;
+	this->xMinion = (float)xMinion;
+	this->yMinion = (float)yMinion;
 }
 
 Bowser::~Bowser()
