+++ Mario/Mario/EventGame.cpp
@@ -113,7 +113,7 @@ void Event::End()
 
 void Event::NewLevel()
 {
-	Window::GetMap()->SetX(newMapX);
+	Window::GetMap()->SetX((float)newMapX);
 	Window::GetMap()->GetPlayer()->SetX(newPlayerX);
 	Window::GetMap()->GetPlayer()->SetY(newPlayerY);
 	Window::GetMap()->SetMoveMap(newMoveMap);
