+++ Mario/Mario/Lakito.cpp
@@ -3,8 +3,8 @@
 
 Lakito::Lakito(int xMinion, int yMinion)
 {
-	this->xMinion = xMinion;
-	this->yMinion = yMinion;
+	this->xMinion = (float)xMinion;
+	this->yMinion = (float)yMinion;
 }
 
 Lakito::~Lakito()
